Extract settings view
